[#3] Collect methods for renaming
