+++ test/test_manipulate_curvature.py
@@ -11,7 +11,7 @@ import numpy as np
 from .fixtures import common_input
 from morphman import manipulate_curvature
 from morphman.common import get_path_names, extract_single_line, \
-                            read_polydata, get_vtk_point_locator, discrete_geometry
+                            read_polydata, vtk_point_locator, discrete_geometry
 
 
 @pytest.mark.parametrize("smooth_line", [True, False])
@@ -40,7 +40,7 @@ def test_decrease_curvature(common_input, smooth_line):
 
     old_centerlines_path = base_path + "_centerline.vtp"
     old_centerlines = read_polydata(old_centerlines_path)
-    old_locator = get_vtk_point_locator(extract_single_line(old_centerlines, 0))
+    old_locator = vtk_point_locator(extract_single_line(old_centerlines, 0))
     old_id1 = old_locator.FindClosestPoint(p1)
     old_id2 = old_locator.FindClosestPoint(p2)
     old_centerline = extract_single_line(old_centerlines, 0, startID=old_id1,
@@ -49,7 +49,7 @@ def test_decrease_curvature(common_input, smooth_line):
     direction = "smoothed" if smooth_line else "extended"
     new_centerlines_path = base_path + "_centerline_new_%s.vtp" % direction
     new_centerlines = read_polydata(new_centerlines_path)
-    new_locator = get_vtk_point_locator(extract_single_line(new_centerlines, 0))
+    new_locator = vtk_point_locator(extract_single_line(new_centerlines, 0))
     new_id1 = new_locator.FindClosestPoint(p1)
     new_id2 = new_locator.FindClosestPoint(p2)
     new_centerline = extract_single_line(new_centerlines, 0, startID=new_id1,
